3rd task is half done

+++ Task3/Program.cs
@@ -13,19 +13,12 @@ namespace Task3
             //List<int> people = new List<int>();
 
             //AddPeople(10, people);
-            ////Console.WriteLine(people.Count);
-            ////people.RemoveAt(0);
-            ////Console.WriteLine(people.Count);
-            ////people.RemoveAt(0);
-            ////Console.WriteLine(people[0]);
-
             //Console.WriteLine("Before removing: ");
             //PrintList(people);
-
             //RemovePeople(people);
 
-            //Console.WriteLine("After removing: ");
-            //PrintList(people);
+
+
 
             //WordFrequency("Вот дом, Который построил Джек. " +
             //    "А это пшеница, Которая в тёмном чулане хранится " +
@@ -33,13 +26,26 @@ namespace Task3
             //    "Которая часто ворует пшеницу, Которая в тёмном чулане хранится " +
             //    "В доме, Который построил Джек. Вот кот, Который пугает и ловит синицу, " +
             //    "Которая часто ворует пшеницу, Которая в тёмном чулане хранится");
+
             Stack<int> stack = new Stack<int>();
             stack.Push(1);
             stack.Push(2);
             stack.Push(3);
 
             DynamicArray<int> test = new DynamicArray<int>(stack);
-            test[4] = 4;
+            Console.WriteLine("Начальный Capacity: " + test.Capacity);
+            Console.WriteLine("Начальный Length: " + test.Length);
+            test.Add(1);
+            Console.WriteLine("Length: " + test.Length);
+            Console.WriteLine("Capacity: " + test.Capacity);
+            test.Add(5);
+            test.Add(1);
+            test.Add(1);
+            Console.WriteLine("Length: " + test.Length);
+            Console.WriteLine("Capacity: " + test.Capacity);
+            test.AddRange(new int[100]);
+            Console.WriteLine("Length: " + test.Length);
+            Console.WriteLine("Capacity: " + test.Capacity);
 
 
 
@@ -51,23 +57,38 @@ namespace Task3
 
         public static void AddPeople(int n, List<int> list)
         {
+            if (n <= 0)
+                throw new ArgumentException("n must be > 0", "n");
+
             for (int i = 0; i < n; i++)
             {
                 list.Add(i + 1);
             }
         }
 
-        public static void RemovePeople(List<int> list)
+        public static void RemovePeople(List<int> list) 
         {
-            while (!(list.Count != 1))
+            if (list.Count < 1)
+                throw new ArgumentException("List<T> must contain at least two elements");
+
+            int Count = 0;
+
+            while (list.Count != 1)
             {
                 int pos = 0;
+
                 for (int i = 0; i < list.Count; i += 2, pos++)
                 {
                     list[pos] = list[i];
                 }
+
                 list.RemoveRange(pos, list.Count - pos);
-            }
+
+                Count++;
+
+                Console.WriteLine("After {0} removing: ", Count);
+                PrintList(list);
+            }    
         }
 
         public static void PrintList(List<int> list)
@@ -88,15 +109,15 @@ namespace Task3
 
             for (int i = 0; i < words.Length; i++)
             {
-                int count = 0;
+                int Count = 0;
 
                 for (int j = 0; j < words.Length; j++)
                 {
-                    if (words[i] == words[j]) count++;
+                    if (words[i] == words[j]) Count++;
                 }
             
                 if(!dict.ContainsKey(words[i])) //проверяем, не находили ли это слово ранее
-                    dict.Add(words[i], count);
+                    dict.Add(words[i], Count);
             }
 
             PrintDictionary<string, int>(dict);
